@@ -12,18 +12,20 @@
     "lint": "yarn lint:js"
   },
   "dependencies": {
-    "core-js": "^3.9.1",
-    "nuxt": "^2.15.3",
-    "storyblok-nuxt": "^1.3.0"
+    "@vueuse/gesture": "^1.0.0",
+    "core-js": "^3.10.1",
+    "nuxt": "^2.15.4",
+    "storyblok-nuxt": "^1.3.1"
   },
   "devDependencies": {
     "@nuxtjs/eslint-config": "^6.0.0",
     "@nuxtjs/eslint-module": "^3.0.2",
-    "@nuxtjs/tailwindcss": "^4.0.1",
+    "@nuxtjs/tailwindcss": "^4.0.3",
     "babel-eslint": "^10.1.0",
-    "eslint": "^7.22.0",
+    "eslint": "^7.24.0",
     "eslint-plugin-nuxt": "^2.0.0",
-    "eslint-plugin-vue": "^7.7.0",
+    "eslint-plugin-vue": "^7.9.0",
+    "nuxt-use-motion": "^1.5.0",
     "nuxt-vite": "^0.0.36"
   },
   "prettier": {
