@@ -2,19 +2,26 @@ module.exports = {
   root: true,
   env: {
     browser: true,
-    node: true
+    node: true,
   },
   parserOptions: {
-    parser: "babel-eslint"
+    parser: "babel-eslint",
   },
   extends: ["@nuxtjs", "plugin:nuxt/recommended"],
   plugins: [],
   rules: {
     quotes: ["error", "double"],
     semi: ["error", "always"],
-    "space-before-function-paren": ["error", "never"],
+    "space-before-function-paren": [
+      "error",
+      {
+        anonymous: "always",
+        named: "never",
+        asyncArrow: "always",
+      },
+    ],
     "comma-dangle": ["error", "only-multiline"],
     "no-console": "off",
-    "no-multiple-empty-lines": "off"
-  }
+    "no-multiple-empty-lines": "off",
+  },
 };
